apps/processor/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
anthropic==0.28.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.7.3
